Use a sidebar placeholder and load content with javascript.
When a reddit page is loaded the side bar contains placeholders for
recent comments and posts.

If javascript is off then a link may be followed and if javascript is
on then the content of the side box will be replaced with the response
from an ajax request.

This prevents google indexing rapidly changing sidebar content.