Fix error shown on production where a vote doesn't have an attribute.
Appears that there are some old votes in the production system where
the vote doesn't have a valid_thing attribute.

This causes an error as when an old vote is changed it accesses
the valid_thing attribute.

The valid_thing will always be true unless the vote is passed spam so
if it doesn't have the attribute we assume that valid_thing is
true. Regardless after the access the value is just passed to
update_score which ignores it.