Reinstate the message inbox.
Added the grey / red envelope icon in the user info sidebar to
indicate new messages / comments and provide a link for users to
access their inbox.

The inbox and sent pages now have the standard sidebar.

The inbox and sent pages now styled in a similar manner to the
comments.