initial implentation of polling per issue 201 this is a fairly hacky implementation and not complete, settings etc. are needed
the basic idea is that if an article is in draft, you make one comment per
poll option. For comments user X makes to user X's posts that are still in
X-drafts, an additional button titled "detach" appears. This takes the post
and changes the author to a system user (should be in config) so people can vote
without affecting user karma. Comments can only be detached if they are on
the user's own drafts to avoid giving a tool to trolls.

UI issues like separating the polls out visually and showing totals aren't done.